revert changes to test/index.js
